refactor: use duration as RetryAfter field type
